Apply renaming of project to core classes
